refactor(tables): split tables into separate views to keep main app clean
