+++ src/ui/Table.tsx
@@ -2,10 +2,11 @@ import { Grid } from "gridjs-react";
 import { OneDArray, TColumn } from "gridjs/dist/src/types";
 import { VFC } from "react";
 import "./Table.css";
-
-type ArrayType = string | number;
+import { ArrayType } from "../utils/CommonTypes";
 
 type ComponentChild = string | number | bigint | boolean | null | undefined;
+
+type TableContentType = string | number;
 
 type Props = {
   data: ArrayType[][];
@@ -33,4 +34,5 @@ const Table: VFC<Props> = ({ data, columns }) => {
     </div>
   );
 };
+export type { TableContentType, ArrayType };
 export default Table;
