+++ src/utils/CommonTypes.tsx
@@ -1,0 +1,7 @@
+type ArrayType = string | number;
+
+type DataProps = {
+  data: ArrayType[][];
+};
+
+export type { DataProps, ArrayType };
