+++ src/utils/Functions.tsx
@@ -1,0 +1,8 @@
+function a11yProps(index: number): { id: string; "aria-controls": string } {
+  return {
+    id: `simple-tab-${index}`,
+    "aria-controls": `simple-tabpanel-${index}`,
+  };
+}
+
+export { a11yProps };
